Fix tests
- Renamed tests folder.
- Upgraded tests.
