--- test/unit/Exception/CouldNotTransitionExceptionInterfaceTest.php
@@ -1,79 +1,0 @@
-<?php
-
-namespace Dhii\State\Exception\UnitTest;
-
-use Xpmock\TestCase;
-use Dhii\State\Exception\CouldNotTransitionExceptionInterface as TestSubject;
-
-/**
- * Tests {@see TestSubject}.
- *
- * @since [*next-version*]
- */
-class CouldNotTransitionExceptionInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\State\Exception\CouldNotTransitionExceptionInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return TestSubject
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getStateMachine()
-            ->getTransition()
-            ->getMessage()
-            ->getCode()
-            ->getPrevious()
-            ->getLine()
-            ->getFile()
-            ->getTrace()
-            ->getTraceAsString()
-            ->__toString();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME,
-            $subject,
-            'A valid instance of the test subject could not be created.'
-        );
-
-        $this->assertInstanceOf(
-            'Dhii\Exception\ThrowableInterface',
-            $subject,
-            'Test subject does not implement parent interface.'
-        );
-
-        $this->assertInstanceOf(
-            'Dhii\State\Exception\StateMachineExceptionInterface',
-            $subject,
-            'Test subject does not implement parent interface.'
-        );
-
-        $this->assertInstanceOf(
-            'Dhii\State\TransitionAwareInterface',
-            $subject,
-            'Test subject does not implement parent interface.'
-        );
-    }
-}
