--- test/unit/StateMachineAwareInterfaceTest.php
@@ -1,52 +1,0 @@
-<?php
-
-namespace Dhii\State\UnitTest;
-
-use Xpmock\TestCase;
-use Dhii\State\StateMachineAwareInterface as TestSubject;
-
-/**
- * Tests {@see TestSubject}.
- *
- * @since [*next-version*]
- */
-class StateMachineAwareInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\State\StateMachineAwareInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return TestSubject
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getStateMachine();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME,
-            $subject,
-            'A valid instance of the test subject could not be created.'
-        );
-    }
-}
