--- test/unit/StateMachineFactoryInterfaceTest.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace Dhii\State\UnitTest;
-
-use Xpmock\TestCase;
-use Dhii\State\StateMachineFactoryInterface as TestSubject;
-
-/**
- * Tests {@see TestSubject}.
- *
- * @since [*next-version*]
- */
-class StateMachineFactoryInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\State\StateMachineFactoryInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return TestSubject
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-                     ->make();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME,
-            $subject,
-            'A valid instance of the test subject could not be created.'
-        );
-
-        $this->assertInstanceOf(
-            'Dhii\Factory\FactoryInterface',
-            $subject,
-            'Test subject does not implement expected interface.'
-        );
-    }
-}
