--- test/unit/StateMachineInterfaceTest.php
@@ -1,53 +1,0 @@
-<?php
-
-namespace Dhii\State\UnitTest;
-
-use Dhii\State\StateMachineInterface as TestSubject;
-use Xpmock\TestCase;
-
-/**
- * Tests {@see TestSubject}.
- *
- * @since [*next-version*]
- */
-class StateMachineInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\State\StateMachineInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return TestSubject
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-                     ->transition()
-                     ->canTransition();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME,
-            $subject,
-            'A valid instance of the test subject could not be created.'
-        );
-    }
-}
